Concatenate window name into WM_CLASS
If you visit https://github.com/mooz/xkeysnail on Google
Chrome, xkeysnail will report WM_CLASS as follows:
  WM_CLASS 'Google-chrome' | active keymaps = ...

After applying this patch, it will be:
  WM_CLASS 'Google-chrome' (mooz/xkeysnail: Yet another
  keyboard remapping tool for X environment - Google
  Chrome)' | active keymaps = ...

This means we can change the keymaps according to the
window title.  This is especially useful when you want to
change the keymap for each tab of the browser.

This patch, however, forces changes to the user's config.py

If possible, I wanted to add the window title as the second
parameter of the lambda function given to "define_keymap",
but the current xkeysnail doesn't support it.

It looks like there's a change that adds a device name to
the second parameter, but It doesn't seem to be working
well.